requirements.txt
numpy
scipy
pillow
opencv-python
pandas
pyyaml
tqdm
einops
albumentations
matplotlib
torch
torchvision
